+++ rcjaRegistration/templates/common/loggedInbase.html
@@ -65,7 +65,7 @@
             </span>
                 <i class="dropdown icon"></i>
                 <div class="menu">
-                    <a href="{% url 'users:profile' %}" class="item">My details</a>
+                    <a href="{% url 'users:details' %}" class="item">My details</a>
                     {% if user.currentlySelectedSchool %}<a href="{% url 'schools:details' %}" class="item">School details</a>{% endif %}
                     <a href = "{% url 'password_change' %}" class="item">Change password</a>
                     <a href = "{% url 'logout' %}" class="item">Log out</a>
@@ -131,7 +131,7 @@
                 </span>
                     <i class="dropdown icon"></i>
                     <div class="menu">
-                        <a href="{% url 'users:profile' %}" class="item">My details</a>
+                        <a href="{% url 'users:details' %}" class="item">My details</a>
                         {% if user.currentlySelectedSchool %}<a href="{% url 'schools:details' %}" class="item">School details</a>{% endif %}
                         <a href = "{% url 'password_change' %}" class="item">Change password</a>
                         <a href = "{% url 'logout' %}" class="item">Log out</a>
