+++ rcjaRegistration/users/urls.py
@@ -3,5 +3,6 @@ from django.urls import path
 from . import views
 app_name = 'users'
 urlpatterns = [
-    path('accounts/profile',views.editProfile, name="profile"),
+    path('accounts/profile',views.details, name="details"),
+    path('accounts/signup', views.signup, name="signup"),
 ]
