+++ rcjaRegistration/users/views.py
@@ -3,17 +3,18 @@ from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.template import loader
 from django.contrib.auth import authenticate, login
-from .forms import UserEditForm
+from .forms import UserForm, UserSignupForm
 from django.http import JsonResponse
 from django.http import HttpResponseForbidden
 from django.forms import modelformset_factory, inlineformset_factory
+from django.urls import reverse
 
 from .models import User
 from userquestions.models import Question, QuestionResponse
 from userquestions.forms import QuestionResponseForm
 
 @login_required
-def editProfile(request):
+def details(request):
     # Questions formset
     questions = Question.objects.all()
     numberQuestions = questions.count()
@@ -36,7 +37,7 @@ def editProfile(request):
         })
 
     if request.method == 'POST':
-        form = UserEditForm(request.POST, instance=request.user)
+        form = UserForm(request.POST, instance=request.user)
         questionFormset = QuestionReponseFormSet(request.POST, instance=request.user, initial=questionResponseInitials)
 
         if form.is_valid() and questionFormset.is_valid():
@@ -48,6 +49,29 @@ def editProfile(request):
 
             return redirect('/')
     else:
-        form = UserEditForm(instance=request.user)
+        form = UserForm(instance=request.user)
         questionFormset = QuestionReponseFormSet(instance=request.user, initial=questionResponseInitials)
     return render(request, 'registration/profile.html', {'form': form, 'questionFormset': questionFormset})
+
+def signup(request):
+    if request.method == 'POST':
+        form = UserSignupForm(request.POST)
+
+        if form.is_valid():
+            # Save user
+            user = form.save()
+            user.set_password(form.cleaned_data["password"])
+            user.save()
+
+            # Save school administrator
+            # if form.cleaned_data['school']:
+            #     SchoolAdministrator.objects.create(user=user, school=form.cleaned_data['school'])
+
+            # Login and redirect
+            login(request, user)
+            return redirect(reverse('users:details'))
+    else:
+        form = UserSignupForm()
+
+    return render(request, 'registration/signup.html', {'form': form})
+
